# docs/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BenchConsole.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/runs": {
      "get": {
        "tags": [
          "Runs"
        ],
        "parameters": [
          {
            "name": "bench",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "verdict",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "plan",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/runs/{id}": {
      "get": {
        "tags": [
          "Runs"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/test-cases": {
      "get": {
        "tags": [
          "TestCases"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GoiTestCaseDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GoiTestCaseDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GoiTestCaseDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TestCases"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "File": {
                    "type": "string",
                    "format": "binary"
                  },
                  "Ten": {
                    "type": "string"
                  },
                  "NguoiTaiLen": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "File": {
                  "style": "form"
                },
                "Ten": {
                  "style": "form"
                },
                "NguoiTaiLen": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GoiTestCaseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoiTestCaseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoiTestCaseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/test-cases/{id}": {
      "get": {
        "tags": [
          "TestCases"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GoiTestCaseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoiTestCaseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoiTestCaseDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TestCases"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/test-cases/{id}/download": {
      "get": {
        "tags": [
          "TestCases"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "GoiTestCaseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "ten": {
            "type": "string",
            "nullable": true
          },
          "tenFileGoc": {
            "type": "string",
            "nullable": true
          },
          "sha256": {
            "type": "string",
            "nullable": true
          },
          "kichThuoc": {
            "type": "integer",
            "format": "int64"
          },
          "soTestCase": {
            "type": "integer",
            "format": "int32"
          },
          "nguoiTaiLen": {
            "type": "string",
            "nullable": true
          },
          "taiLenLuc": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}